refactor: rename textbook selectors file
